More incorrect screenshot JNI code.
Not sure why we aren't getting warnings but these casts are to the wrong
type.

Bug: 34758075
Test: Manual

+++ core/jni/android_view_SurfaceControl.cpp
@@ -260,7 +260,7 @@ static void nativeScreenshot(JNIEnv* env, jclass clazz, jobject displayTokenObj,
             }
             ScreenshotClient::capture(displayToken,
                     consumer->getIGraphicBufferProducer(), sourceCrop,
-                    width, height, uint32_t(minLayer), uint32_t(maxLayer),
+                    width, height, minLayer, maxLayer,
                     useIdentityTransform);
         }
     }
